Create IDCA and manage encapsulation
