+++ backend/Procfile
@@ -1,0 +1,1 @@
+web: uvicorn ingestion_agent.main:app --host 0.0.0.0 --port=8080
