+++ frontend/src/App.jsx
@@ -2,9 +2,6 @@ import { useState } from "react";
 
 export default function App() {
   const [file, setFile] = useState(null);
-  const [gcsURI, setGcsURI] = useState("");
-  const [prompt, setPrompt] = useState("");
-  const [response, setResponse] = useState("");
 
   const upload = async () => {
     console.log(
@@ -27,43 +24,11 @@ export default function App() {
     }
   };
 
-  const ask = async () => {
-    const res = await fetch(process.env.REACT_APP_INGESTION_URL + "/query", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ prompt, gcs_uri: gcsURI }),
-    });
-
-    const data = await res.json();
-    console.log("Query response:", data);
-    setResponse(data.response);
-  };
-
   return (
     <div style={{ padding: 20 }}>
       <h1>Upload Manuscript</h1>
       <input type="file" onChange={(e) => setFile(e.target.files[0])} />
       <button onClick={upload}>Upload</button>
-      {gcsURI && (
-        <>
-          <h2>Ask About the File</h2>
-          <input
-            type="text"
-            value={prompt}
-            onChange={(e) => setPrompt(e.target.value)}
-            placeholder="Ask a question..."
-            style={{ width: "300px" }}
-          />
-          <button onClick={ask}>Ask</button>
-          {response && typeof response === "string" && (
-            <p>
-              <strong>Response:</strong> {response}
-            </p>
-          )}
-        </>
-      )}
     </div>
   );
 }
